Move the lsp decorator into utils in lsp
